backend/requirements.txt
# FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Database & Storage
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-cloud-bigquery==3.22.0
firebase-admin==6.5.0

# AI & ML
google-cloud-aiplatform==1.49.0
langchain==0.2.1
langchain-google-genai==1.0.5
langchain-community==0.2.1

# Bias Detection & ML
aif360==0.6.1
scikit-learn==1.4.2
onnxruntime==1.17.3
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0

# Explainability
lime==0.2.0.1
shap==0.45.0

# Synthetic Data
sdv==1.12.1

# Async Task Queue
celery[redis]==5.4.0
redis==5.0.4

# Auth
PyJWT==2.8.0
cryptography==42.0.5

# PDF & Badge Generation
reportlab==4.2.0
qrcode[pil]==7.4.2
svgwrite==1.4.3

# PII Detection
python-stdnum==1.20

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
orjson==3.10.3
python-dateutil==2.9.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
